requirement.txt
python-docx
comtypes